11ПР

+++ src/utils/Api.js
@@ -56,19 +56,37 @@ class Api {
     }).then((res) => this._checkError(res));
   };
 
-  likeCard(cardId) {
-    return fetch(`${this._url}/cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    }).then((res) => this._checkError(res));
-  };
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._url}/cards/${cardId}/likes`,
+        {
+          method: 'PUT',
+          headers: this._headers,
+        })
+        .then((res) => this._checkError(res));
+    } else {
+      return fetch(`${this._url}/cards/${cardId}/likes`,
+        {
+          method: 'DELETE',
+          headers: this._headers,
+        })
+        .then((res) => this._checkError(res));
+    }
+  }
 
-  unlikeCard(cardId) {
-    return fetch(`${this._url}/cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    }).then((res) => this._checkError(res));
-  };
+  // likeCard(cardId) {
+  //   return fetch(`${this._url}/cards/${cardId}/likes`, {
+  //     method: "PUT",
+  //     headers: this._headers,
+  //   }).then((res) => this._checkError(res));
+  // };
+
+  // unlikeCard(cardId) {
+  //   return fetch(`${this._url}/cards/${cardId}/likes`, {
+  //     method: "DELETE",
+  //     headers: this._headers,
+  //   }).then((res) => this._checkError(res));
+  // };
 }
 
 const apiConfig = {
